+++ test/external/fuzz_schedule.py
@@ -1,8 +1,7 @@
 from collections import defaultdict
-from typing import DefaultDict, Dict, List
+from typing import DefaultDict, List, Set, TypeVar, Dict
 from tinygrad.buffer import Buffer
 from tinygrad.engine.realize import run_schedule
-from tinygrad.features.toposort import find_all_sorts
 from tinygrad.helpers import DEBUG, colored
 from tinygrad.lazy import LazyBuffer
 from tinygrad.engine.schedule import graph_schedule
@@ -31,3 +30,31 @@ def fuzz_schedule(outs: List[LazyBuffer]):
     raw_outs = [[out.as_buffer().tobytes() for out in si.outputs] for si in items]
     assert all(o == raw_outs[0] for o in raw_outs)
   if DEBUG >= 2: print(colored("all toposorts passed", "green"))
+
+
+T = TypeVar("T")
+def find_all_sorts(graph:DefaultDict[T, List[T]], in_degree:DefaultDict[T, int]) -> List[List[T]]:
+  visited: Set[T] = set()
+  ret: List[List[T]] = []
+  path: List[T] = []
+
+  def recurse_paths(path:List[T]):
+    for v, d in in_degree.items():
+      if d != 0 or v in visited: continue
+      for u in graph[v]: in_degree[u] -= 1
+      path.append(v)
+      visited.add(v)
+      recurse_paths(path)
+      # backtrack
+      for u in graph[v]: in_degree[u] += 1
+      path.pop()
+      visited.remove(v)
+    if len(path) == len(in_degree): ret.append([*path])
+  recurse_paths(path)
+
+  if len(ret) == 0: raise RuntimeError("detected cycle in the graph")
+  # verify all paths are unique
+  assert len(ret) == len(set(map(tuple, ret)))
+  # backtrack cleanup
+  for v in in_degree: in_degree[v] = 0
+  return ret
