--- tinygrad/features/toposort.py
@@ -1,28 +1,0 @@
-from typing import DefaultDict, List, Set, TypeVar
-
-T = TypeVar("T")
-def find_all_sorts(graph:DefaultDict[T, List[T]], in_degree:DefaultDict[T, int]) -> List[List[T]]:
-  visited: Set[T] = set()
-  ret: List[List[T]] = []
-  path: List[T] = []
-
-  def recurse_paths(path:List[T]):
-    for v, d in in_degree.items():
-      if d != 0 or v in visited: continue
-      for u in graph[v]: in_degree[u] -= 1
-      path.append(v)
-      visited.add(v)
-      recurse_paths(path)
-      # backtrack
-      for u in graph[v]: in_degree[u] += 1
-      path.pop()
-      visited.remove(v)
-    if len(path) == len(in_degree): ret.append([*path])
-  recurse_paths(path)
-
-  if len(ret) == 0: raise RuntimeError("detected cycle in the graph")
-  # verify all paths are unique
-  assert len(ret) == len(set(map(tuple, ret)))
-  # backtrack cleanup
-  for v in in_degree: in_degree[v] = 0
-  return ret
